project_bill/requirements.txt
django
djangorestframework
psycopg2
gunicorn
